Works for mid size data sets

+++ code/Point2D.hpp
@@ -31,7 +31,7 @@ using namespace std;
 
 namespace geometry {
   // Defines the implementation and precision of coordinates
-  typedef float coord_t;
+  typedef int coord_t;
   
   /////////////////////////////////////////////////////////////////////////////
   // Point2D interface                                                       //
